Apply clang-tidy performance-enum-size
Always specify the underlying type for an enum. Without this the
compiler is free to choose the type (with some constraints). Gcc and
clang seem to pick 'int' or 'unsigned'.

In this patch we instead choose 'uint8_t', unless there's a reason to
pick something else. Usually this difference doesn't matter (e.g. when
passing the enum as a function parameter). But it may make a difference
when that enum is stored as a member variable, then it will only take 1
instead of 4 bytes (and possibly reduce padding because of alignment
constraints).

+++ src/cpu/CPUCore.cc
@@ -214,8 +214,8 @@
 
 namespace openmsx {
 
-enum Reg8  : int { A, F, B, C, D, E, H, L, IXH, IXL, IYH, IYL, REG_I, REG_R, DUMMY };
-enum Reg16 : int { AF, BC, DE, HL, IX, IY, SP };
+enum Reg8  : uint8_t { A, F, B, C, D, E, H, L, IXH, IXL, IYH, IYL, REG_I, REG_R, DUMMY };
+enum Reg16 : uint8_t { AF, BC, DE, HL, IX, IY, SP };
 
 // flag positions
 static constexpr byte S_FLAG = 0x80;
